Move the BridgeCloudNetHelper methods to the BridgeHelper
